Move translations to @aws-amplify/ui
