+++ packages/angular/projects/ui-angular/src/lib/components/amplify-sign-in/amplify-sign-in.component.ts
@@ -13,7 +13,7 @@ import { StateMachineService } from '../../services/state-machine.service';
 import { AuthPropService } from '../../services/authenticator-context.service';
 import { Subscription } from 'xstate';
 import { AuthMachineState, getActorState, SignInState } from '@aws-amplify/ui';
-import { DefaultTexts, translate } from '../../common';
+import { DefaultTexts, translate } from '@aws-amplify/ui';
 
 const logger = new Logger('SignIn');
